Unify help command, localize it.
