+++ Resources/Locale/en-US/commands.ftl
@@ -1,0 +1,15 @@
+﻿### Localization for engine console commands
+
+## help command
+
+cmd-help-desc = Display general help or help text for a specific command
+cmd-help-help = Usage: help [command name]
+    When no command name is provided, displays general-purpose help text. If a command name is provided, displays help text for that command.
+
+cmd-help-no-args = To display help for a specific command, write 'help <command>'. To list all available commands, write 'list'.
+cmd-help-unknown = Unknown command: { $command }
+cmd-help-top = { $command } - { $description }
+cmd-help-invalid-args = Invalid amount of arguments.
+cmd-help-arg-cmdname = [command name]
+
+##
